+++ CHANGELOG.md
@@ -9,10 +9,10 @@
   
 ## Changelog
   
+### **HEAD -> main** 2021/09/09 mandic00@live.com
+
+
 ### **0.3.1** 2021/09/09 mandic00@live.com
-
-
-### **origin/main** 2021/09/08 mandic00@live.com
 
 - stable release