+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @vladmandic/build  
 
-  Version: **0.3.1**  
+  Version: **0.3.2**  
   Description: **Build: Automated CI Platform for NodeJS**  
   
   Author: **Vladimir Mandic <mandic00@live.com>**  
@@ -9,7 +9,11 @@
   
 ## Changelog
   
-### **HEAD -> main** 2021/09/09 mandic00@live.com
+### **0.3.2** 2021/09/09 mandic00@live.com
+
+- added results parser
+
+### **origin/main** 2021/09/09 mandic00@live.com
 
 - added defaults for all config options
 - add types and typedocs